+++ providers/dns/cloudflare/cloudflare_test.go
@@ -1,4 +1,4 @@
-package acme
+package cloudflare
 
 import (
 	"os"
@@ -29,26 +29,26 @@ func restoreCloudFlareEnv() {
 	os.Setenv("CLOUDFLARE_API_KEY", cflareAPIKey)
 }
 
-func TestNewDNSProviderCloudFlareValid(t *testing.T) {
+func TestNewDNSProviderValid(t *testing.T) {
 	os.Setenv("CLOUDFLARE_EMAIL", "")
 	os.Setenv("CLOUDFLARE_API_KEY", "")
-	_, err := NewDNSProviderCloudFlare("123", "123")
+	_, err := NewDNSProvider("123", "123")
 	assert.NoError(t, err)
 	restoreCloudFlareEnv()
 }
 
-func TestNewDNSProviderCloudFlareValidEnv(t *testing.T) {
+func TestNewDNSProviderValidEnv(t *testing.T) {
 	os.Setenv("CLOUDFLARE_EMAIL", "test@example.com")
 	os.Setenv("CLOUDFLARE_API_KEY", "123")
-	_, err := NewDNSProviderCloudFlare("", "")
+	_, err := NewDNSProvider("", "")
 	assert.NoError(t, err)
 	restoreCloudFlareEnv()
 }
 
-func TestNewDNSProviderCloudFlareMissingCredErr(t *testing.T) {
+func TestNewDNSProviderMissingCredErr(t *testing.T) {
 	os.Setenv("CLOUDFLARE_EMAIL", "")
 	os.Setenv("CLOUDFLARE_API_KEY", "")
-	_, err := NewDNSProviderCloudFlare("", "")
+	_, err := NewDNSProvider("", "")
 	assert.EqualError(t, err, "CloudFlare credentials missing")
 	restoreCloudFlareEnv()
 }
@@ -58,7 +58,7 @@ func TestCloudFlarePresent(t *testing.T) {
 		t.Skip("skipping live test")
 	}
 
-	provider, err := NewDNSProviderCloudFlare(cflareEmail, cflareAPIKey)
+	provider, err := NewDNSProvider(cflareEmail, cflareAPIKey)
 	assert.NoError(t, err)
 
 	err = provider.Present(cflareDomain, "", "123d==")
@@ -70,9 +70,9 @@ func TestCloudFlareCleanUp(t *testing.T) {
 		t.Skip("skipping live test")
 	}
 
-	time.Sleep(time.Second * 1)
+	time.Sleep(time.Second * 2)
 
-	provider, err := NewDNSProviderCloudFlare(cflareEmail, cflareAPIKey)
+	provider, err := NewDNSProvider(cflareEmail, cflareAPIKey)
 	assert.NoError(t, err)
 
 	err = provider.CleanUp(cflareDomain, "", "123d==")